refactor: extract proxy session helpers
