+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - 非流式 collect/retry 路径从 `proxy-handler.ts` 抽到独立 `non-streaming-handler.ts`，并把 `buildCodexApi`、image generation usage 标记、Codex error prefix 清理收敛到 `proxy-handler-utils.ts`，降低主 handler 对 empty-response retry / premature-close / affinity 逻辑的耦合；新增模块边界测试、collect 阶段 `previous_response_not_found` strip-retry guard、non-stream affinity metadata guard（`src/routes/shared/non-streaming-handler.ts`、`src/routes/shared/proxy-handler-utils.ts`、`src/routes/shared/proxy-handler.ts`、`tests/unit/routes/shared/non-streaming-handler-boundary.test.ts`、`tests/integration/proxy-handler.test.ts`）
 - `proxy-handler.ts` 的共享类型契约抽到独立 `proxy-handler-types.ts`，route / helper / test helper 统一从类型模块导入 `ProxyRequest`、`FormatAdapter`、translator options 与 handler options，避免 `response-processor.ts` / `non-streaming-handler.ts` 为类型依赖反向指向运行时 orchestrator；新增 structural guard 锁住完整契约导出、handler 不再声明共享接口、以及禁止从 `proxy-handler.js` 混入这些共享类型（`src/routes/shared/proxy-handler-types.ts`、`src/routes/shared/proxy-handler.ts`、`src/routes/shared/response-processor.ts`、`src/routes/shared/non-streaming-handler.ts`、`tests/unit/routes/shared/proxy-handler-type-boundary.test.ts`）
 - API-key direct upstream 路径从 `proxy-handler.ts` 抽到独立 `direct-request-handler.ts`，route 层直接导入 direct handler，`proxy-handler.ts` 专注账号池 acquire / retry / release 编排；流式错误 SSE helper 抽到 `stream-error-response.ts` 供 Codex proxy 与 direct upstream 共用，并新增 structural guard 防止 `handleDirectRequest` 回流到 `proxy-handler.js`（`src/routes/shared/direct-request-handler.ts`、`src/routes/shared/stream-error-response.ts`、`src/routes/shared/proxy-handler.ts`、`src/routes/chat.ts`、`src/routes/messages.ts`、`src/routes/gemini.ts`、`src/routes/responses.ts`、`tests/unit/routes/shared/direct-request-handler-boundary.test.ts`）
+- prompt-cache identity、implicit-resume 判定、continuation input / function-call-output helper 从 `proxy-handler.ts` 抽到独立 `proxy-session-helpers.ts`，让主 handler 只保留编排与状态变更；原 implicit-resume 单测改为直接覆盖 helper 模块，并新增 structural guard 防止这些 session helper 回流到 `proxy-handler.js`（`src/routes/shared/proxy-session-helpers.ts`、`src/routes/shared/proxy-handler.ts`、`tests/unit/routes/shared/proxy-handler-implicit-resume.test.ts`、`tests/unit/routes/shared/proxy-session-helpers-boundary.test.ts`）
 - 更新提示默认不再弹窗：新增 `update.show_update_dialog`（默认 `false`），Dashboard 设置页可手动开启“显示更新弹窗”。Web 自更新弹窗与 Electron 自动更新的系统对话框都受该开关控制；更新检查和托盘/菜单入口仍保留，避免默认后台检查打断使用
 - `src/routes/api-keys.ts` 简化第三方 API key 绑定路由：合并 add/import 的重复 Zod schema，统一 JSON 解析与校验错误返回，复用按 `models` 展开写入的逻辑，并新增 `tests/unit/routes/api-keys.test.ts` 覆盖添加、导入、导出、custom provider 校验与批量删除，确保行为不变
 - README / README_EN 的 Codex CLI + Codex Desktop 两节示例从 `env_key = "PROXY_API_KEY"` 改成 `[model_providers.proxy_codex.http_headers]` 内嵌 `Authorization = "Bearer ..."`：原写法在 GUI 客户端启动时会因为 macOS / Windows GUI 进程不继承 shell rc 的环境变量而报 `Missing environment variable: PROXY_API_KEY`，普通用户得额外学 `launchctl setenv` 或 LaunchAgent 才能让 Codex Desktop 看到环境变量；http_headers 把 key 直接写在 config 文件里，重启 Codex 即用。`env_key` 写法作为「需要密钥从配置文件解耦」（多人共享 / 仓库提交）场景的备选保留在文档说明里
